# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Sports Event Platform API
  description: AI 驅動的兒童體育賽事電商平台 API
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/events:
    get:
      tags:
        - Events
      summary: 查詢賽事列表
      description: |
        查詢賽事列表，支援依年齡、日期區間、地點篩選，以及分頁與排序功能。
        此 API 設計為 AI Agent 可呼叫的工具，回傳結構化資料便於 AI 處理。
      operationId: getEvents
      parameters:
        # Filters
        - name: age
          in: query
          description: 依參賽者年齡篩選（會回傳 age_min <= age <= age_max 的賽事）
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 18
            example: 10
        - name: dateFrom
          in: query
          description: 賽事開始日期區間起點（含）
          required: false
          schema:
            type: string
            format: date
            example: "2026-02-01"
        - name: dateTo
          in: query
          description: 賽事開始日期區間終點（含）
          required: false
          schema:
            type: string
            format: date
            example: "2026-03-31"
        - name: location
          in: query
          description: 依地點篩選（模糊比對）
          required: false
          schema:
            type: string
            maxLength: 100
            example: "台北"
        # Pagination
        - name: page
          in: query
          description: 頁碼（從 0 開始）
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: size
          in: query
          description: 每頁筆數
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        # Sorting
        - name: sort
          in: query
          description: 排序欄位與方向
          required: false
          schema:
            type: string
            enum:
              - startTime,asc
              - startTime,desc
            default: "startTime,asc"
            example: "startTime,asc"
      responses:
        "200":
          description: 成功取得賽事列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPageResponse"
              examples:
                success:
                  $ref: "#/components/examples/EventPageResponseExample"
        "400":
          description: 請求參數錯誤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidAge:
                  $ref: "#/components/examples/InvalidAgeError"
                invalidDateRange:
                  $ref: "#/components/examples/InvalidDateRangeError"
        "500":
          description: 伺服器內部錯誤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                serverError:
                  $ref: "#/components/examples/ServerError"
  /api/events/{id}:
    get:
      tags:
        - Events
      summary: 取得賽事詳情
      description: |
        根據賽事 ID 取得完整賽事資訊，包含年齡限制、報名狀態等詳細資料。
        此 API 可供 AI Agent 呼叫，取得特定賽事的完整資訊以回答使用者問題。
      operationId: getEventById
      parameters:
        - name: id
          in: path
          description: 賽事唯一識別碼
          required: true
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: 成功取得賽事詳情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetailResponse"
              examples:
                success:
                  $ref: "#/components/examples/EventDetailResponseExample"
        "400":
          description: 請求參數錯誤（ID 格式不正確）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidId:
                  $ref: "#/components/examples/InvalidIdError"
        "404":
          description: 賽事不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  $ref: "#/components/examples/EventNotFoundError"
        "500":
          description: 伺服器內部錯誤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                serverError:
                  $ref: "#/components/examples/ServerError"
  /api/orders:
    post:
      tags:
        - Orders
      summary: 建立報名訂單
      description: |
        為指定賽事建立報名訂單。

        **驗證順序（依序執行，任一失敗即停止）：**
        1. 請求欄位格式驗證（必填欄位、長度限制、格式規範）
        2. `child.birthDate` 必須是過去的日期（不可為今日或未來日期）
        3. 賽事必須存在（`eventId` 對應到有效賽事）
        4. 賽事報名狀態必須為 OPEN，且當前時間在 `registrationDeadline` 之前
        5. 賽事必須尚有名額（`registeredCount < capacity`）
        6. 小孩於賽事當日的年齡須符合年齡限制（見下方計算公式）
        7. 同一小孩不可重複報名同一賽事（依 `child.name + child.birthDate` 判斷）

        **年齡計算公式：**
        ```
        childAgeAtEvent = 賽事開始日期(event.startTime) 與出生日期(child.birthDate) 之間的完整年數
        ```
        範例：birthDate = 2016-05-15，event.startTime = 2026-03-15
        → childAgeAtEvent = 9（三月尚未滿 10 歲生日）

        **嚴格 vs. 彈性年齡驗證（由 event.strictAgeEnforcement 控制）：**
        - `strictAgeEnforcement = true`（預設）：年齡不在範圍內直接拒絕（HTTP 409, AGE_NOT_ELIGIBLE）
        - `strictAgeEnforcement = false`：年齡不在範圍內仍允許報名，但回應中加入 `warnings` 陣列提示

        **付款期限計算：**
        ```
        paymentDeadline = min(event.registrationDeadline, 訂單建立時間 + 72小時)
        ```

        **訂單狀態生命週期：**
        ```
        PENDING ─付款完成─→ CONFIRMED ─賽事結束─→ COMPLETED
           │                     │
           │（使用者取消/逾期）   │（退款申請通過/賽事取消）
           ↓                     ↓
        CANCELLED             REFUNDED
        ```
        訂單建立後初始狀態為 PENDING，等待付款確認。
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            examples:
              basic:
                $ref: "#/components/examples/CreateOrderRequestExample"
      responses:
        "201":
          description: 訂單建立成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
              examples:
                success:
                  $ref: "#/components/examples/OrderCreatedExample"
                nonStrictAge:
                  $ref: "#/components/examples/OrderNonStrictAgeSuccess"
        "400":
          description: 請求參數驗證失敗
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingField:
                  $ref: "#/components/examples/OrderMissingFieldError"
                invalidEmail:
                  $ref: "#/components/examples/OrderInvalidEmailError"
                invalidPhone:
                  $ref: "#/components/examples/OrderInvalidPhoneError"
                invalidBirthDate:
                  $ref: "#/components/examples/OrderInvalidBirthDateError"
        "404":
          description: 賽事不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                eventNotFound:
                  $ref: "#/components/examples/EventNotFoundError"
        "409":
          description: 業務邏輯衝突
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                ageNotEligible:
                  $ref: "#/components/examples/OrderAgeNotEligibleError"
                eventFull:
                  $ref: "#/components/examples/OrderEventFullError"
                registrationClosed:
                  $ref: "#/components/examples/OrderRegistrationClosedError"
                duplicateRegistration:
                  $ref: "#/components/examples/OrderDuplicateError"
        "500":
          description: 伺服器內部錯誤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                serverError:
                  $ref: "#/components/examples/ServerError"
  /api/admin/events:
    get:
      tags:
        - Admin
      summary: 管理員查詢賽事列表
      description: |
        管理員查看所有賽事列表，包含報名統計等管理資訊。

        **權限要求**: 此 API 需要管理員權限。在實際環境中，需透過 JWT token 或 session 驗證管理員身份。
        在此示範專案中，假設已通過認證機制驗證管理員身份。
      operationId: adminGetEvents
      parameters:
        - name: status
          in: query
          description: 依賽事報名狀態篩選
          required: false
          schema:
            type: string
            enum:
              - OPEN
              - CLOSED
              - FULL
            example: "OPEN"
        - name: dateFrom
          in: query
          description: 賽事開始日期區間起點
          required: false
          schema:
            type: string
            format: date
            example: "2026-02-01"
        - name: dateTo
          in: query
          description: 賽事開始日期區間終點
          required: false
          schema:
            type: string
            format: date
            example: "2026-03-31"
        - name: page
          in: query
          description: 頁碼（從 0 開始）
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: size
          in: query
          description: 每頁筆數
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: sort
          in: query
          description: 排序欄位與方向
          required: false
          schema:
            type: string
            enum:
              - createdAt,desc
              - createdAt,asc
              - startTime,asc
              - startTime,desc
            default: "createdAt,desc"
            example: "createdAt,desc"
      responses:
        "200":
          description: 成功取得賽事列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminEventPageResponse"
              examples:
                success:
                  $ref: "#/components/examples/AdminEventPageResponseExample"
        "401":
          description: 未認證或認證失效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: "#/components/examples/UnauthorizedError"
        "403":
          description: 權限不足（非管理員）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                forbidden:
                  $ref: "#/components/examples/ForbiddenError"
        "500":
          description: 伺服器內部錯誤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                serverError:
                  $ref: "#/components/examples/ServerError"

    post:
      tags:
        - Admin
      summary: 管理員建立賽事
      description: |
        管理員建立新賽事。

        **權限要求**: 此 API 需要管理員權限。
      operationId: adminCreateEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
            examples:
              basic:
                $ref: "#/components/examples/CreateEventRequestExample"
      responses:
        "201":
          description: 賽事建立成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminEventResponse"
              examples:
                success:
                  $ref: "#/components/examples/AdminEventCreatedExample"
        "400":
          description: 請求參數驗證失敗
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validationError:
                  $ref: "#/components/examples/AdminEventValidationError"
        "401":
          description: 未認證或認證失效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: "#/components/examples/UnauthorizedError"
        "403":
          description: 權限不足（非管理員）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                forbidden:
                  $ref: "#/components/examples/ForbiddenError"
        "500":
          description: 伺服器內部錯誤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                serverError:
                  $ref: "#/components/examples/ServerError"
components:
  schemas:
...
    EventDetailResponse:
      type: object
      description: 賽事詳情回應
      properties:
        success:
          type: boolean
          description: 請求是否成功
          example: true
        data:
          $ref: "#/components/schemas/EventDetail"
        timestamp:
          type: string
          format: date-time
          description: 回應時間戳記
          example: "2026-01-26T10:30:00+08:00"
      required:
        - success
        - data
        - timestamp
...
    EventPageResponse:
      type: object
      description: 賽事列表分頁回應
      properties:
        success:
          type: boolean
          description: 請求是否成功
          example: true
        data:
          type: array
          description: 賽事列表
          items:
            $ref: "#/components/schemas/Event"
        page:
          $ref: "#/components/schemas/PageInfo"
        timestamp:
          type: string
          format: date-time
          description: 回應時間戳記
          example: "2026-01-26T10:30:00+08:00"
      required:
        - success
        - data
        - page
        - timestamp

    # ===== Order Schemas =====
    CreateOrderRequest:
      type: object
      description: 建立訂單請求
      properties:
        eventId:
          type: string
          format: uuid
          description: 賽事 ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        parent:
          $ref: "#/components/schemas/ParentInfo"
        child:
          $ref: "#/components/schemas/ChildInfo"
        emergencyContact:
          $ref: "#/components/schemas/EmergencyContact"
        notes:
          type: string
          description: 備註（可選）
          maxLength: 500
          example: "小孩對花生過敏"
      required:
        - eventId
        - parent
        - child
...
    OrderResponse:
      type: object
      description: 訂單回應
      properties:
        success:
          type: boolean
          description: 請求是否成功
          example: true
        data:
          $ref: "#/components/schemas/Order"
        timestamp:
          type: string
          format: date-time
          description: 回應時間戳記
          example: "2026-01-26T10:30:00+08:00"
      required:
        - success
        - data
        - timestamp

    # ===== Admin Schemas =====
...
    AdminEventResponse:
      type: object
      description: 管理員賽事回應
      properties:
        success:
          type: boolean
          description: 請求是否成功
          example: true
        data:
          $ref: "#/components/schemas/AdminEventDetail"
        timestamp:
          type: string
          format: date-time
          description: 回應時間戳記
          example: "2026-01-26T10:30:00+08:00"
      required:
        - success
        - data
        - timestamp
    AdminEventPageResponse:
      type: object
      description: 管理員賽事列表分頁回應
      properties:
        success:
          type: boolean
          description: 請求是否成功
          example: true
        data:
          type: array
          description: 賽事列表
          items:
            $ref: "#/components/schemas/AdminEventDetail"
        page:
          $ref: "#/components/schemas/PageInfo"
        timestamp:
          type: string
          format: date-time
          description: 回應時間戳記
          example: "2026-01-26T10:30:00+08:00"
      required:
        - success
        - data
        - page
        - timestamp

    # ===== Error Schema =====
    ErrorResponse:
      type: object
      description: 錯誤回應
      properties:
        success:
          type: boolean
          description: 請求是否成功
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: 錯誤代碼
              example: "INVALID_PARAMETER"
            message:
              type: string
              description: 錯誤訊息（使用者可讀）
              example: "年齡參數必須介於 1 到 18 之間"
            details:
              type: array
              description: 詳細錯誤資訊（可選）
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: 錯誤欄位
                    example: "age"
                  reason:
                    type: string
                    description: 錯誤原因
                    example: "must be between 1 and 18"
            traceId:
              type: string
              description: 追蹤識別碼（用於除錯）
              example: "abc123def456"
          required:
            - code
            - message
        timestamp:
          type: string
          format: date-time
          description: 回應時間戳記
          example: "2026-01-26T10:30:00+08:00"
      required:
        - success
        - error
        - timestamp
  examples:
    EventPageResponseExample:
      summary: 成功回應範例
      value:
        success: true
        data:
          - id: "550e8400-e29b-41d4-a716-446655440000"
            name: "2026 春季兒童籃球營"
            ageMin: 8
            ageMax: 12
            startTime: "2026-03-15T09:00:00+08:00"
            location: "台北市大安運動中心"
          - id: "550e8400-e29b-41d4-a716-446655440001"
            name: "兒童游泳體驗班"
            ageMin: 6
            ageMax: 10
            startTime: "2026-03-20T14:00:00+08:00"
            location: "台北市信義運動中心"
        page:
          page: 0
          size: 10
          totalElements: 2
          totalPages: 1
          hasNext: false
          hasPrevious: false
        timestamp: "2026-01-26T10:30:00+08:00"
    InvalidAgeError:
      summary: 年齡參數錯誤
      value:
        success: false
        error:
          code: "INVALID_PARAMETER"
          message: "年齡參數必須介於 1 到 18 之間"
          details:
            - field: "age"
              reason: "must be between 1 and 18"
          traceId: "abc123def456"
        timestamp: "2026-01-26T10:30:00+08:00"
    InvalidDateRangeError:
      summary: 日期區間錯誤
      value:
        success: false
        error:
          code: "INVALID_PARAMETER"
          message: "結束日期不可早於開始日期"
          details:
            - field: "dateFrom"
              reason: "dateFrom must be before or equal to dateTo"
            - field: "dateTo"
              reason: "dateTo must be after or equal to dateFrom"
          traceId: "abc123def789"
        timestamp: "2026-01-26T10:30:00+08:00"
    ServerError:
      summary: 伺服器內部錯誤
      value:
        success: false
        error:
          code: "INTERNAL_SERVER_ERROR"
          message: "系統發生非預期錯誤，請稍後再試"
          traceId: "xyz789abc123"
        timestamp: "2026-01-26T10:30:00+08:00"
    EventDetailResponseExample:
      summary: 賽事詳情成功回應
      value:
        success: true
        data:
          id: "550e8400-e29b-41d4-a716-446655440000"
          name: "2026 春季兒童籃球營"
          description: "專為 8-12 歲兒童設計的籃球訓練營，由專業教練指導基礎籃球技巧，包含運球、傳球、投籃等基礎訓練。活動結束後將舉辦小型友誼賽。"
          ageRestriction:
            minAge: 8
            maxAge: 12
            description: "適合 8 至 12 歲兒童參加"
            strictEnforcement: true
          startTime: "2026-03-15T09:00:00+08:00"
          endTime: "2026-03-15T12:00:00+08:00"
          location: "台北市大安運動中心"
          address: "台北市大安區辛亥路三段55號"
          capacity: 30
          registeredCount: 18
          registrationStatus: "OPEN"
          registrationDeadline: "2026-03-10T23:59:59+08:00"
          fee: 1500
          organizer: "台北市籃球協會"
          contactEmail: "events@basketball.org.tw"
          contactPhone: "02-1234-5678"
          createdAt: "2026-01-01T10:00:00+08:00"
          updatedAt: "2026-01-20T15:30:00+08:00"
        timestamp: "2026-01-26T10:30:00+08:00"
    InvalidIdError:
      summary: ID 格式錯誤
      value:
        success: false
        error:
          code: "INVALID_PARAMETER"
          message: "賽事 ID 格式不正確"
          details:
            - field: "id"
              reason: "must be a valid UUID format"
          traceId: "abc123def470"
        timestamp: "2026-01-26T10:30:00+08:00"
    EventNotFoundError:
      summary: 賽事不存在
      value:
        success: false
        error:
          code: "RESOURCE_NOT_FOUND"
          message: "找不到指定的賽事"
          details:
            - field: "id"
              reason: "event with the specified ID does not exist"
          traceId: "abc123def471"
        timestamp: "2026-01-26T10:30:00+08:00"

    # ===== Order Examples =====
    CreateOrderRequestExample:
      summary: 建立訂單請求範例
      value:
        eventId: "550e8400-e29b-41d4-a716-446655440000"
        parent:
          name: "王小明"
          email: "parent@example.com"
          phone: "0912345678"
        child:
          name: "王小華"
          birthDate: "2016-05-15"
          gender: "MALE"
        emergencyContact:
          name: "王大明"
          phone: "0987654321"
          relationship: "父親"
        notes: "小孩對花生過敏"
    OrderCreatedExample:
      summary: 訂單建立成功
      value:
        success: true
        data:
          id: "660e8400-e29b-41d4-a716-446655440000"
          eventId: "550e8400-e29b-41d4-a716-446655440000"
          eventName: "2026 春季兒童籃球營"
          status: "PENDING"
          parent:
            name: "王小明"
            email: "parent@example.com"
            phone: "0912345678"
          child:
            name: "王小華"
            birthDate: "2016-05-15"
            gender: "MALE"
          childAgeAtEvent: 10
          emergencyContact:
            name: "王大明"
            phone: "0987654321"
            relationship: "父親"
          notes: "小孩對花生過敏"
          fee: 1500
          paymentDeadline: "2026-02-01T23:59:59+08:00"
          createdAt: "2026-01-26T10:30:00+08:00"
          updatedAt: "2026-01-26T10:30:00+08:00"
        timestamp: "2026-01-26T10:30:00+08:00"
    OrderMissingFieldError:
      summary: 缺少必填欄位
      value:
        success: false
        error:
          code: "VALIDATION_ERROR"
          message: "請求參數驗證失敗"
          details:
            - field: "parent.name"
              reason: "must not be blank"
            - field: "child.birthDate"
              reason: "must not be null"
          traceId: "ord123def001"
        timestamp: "2026-01-26T10:30:00+08:00"
    OrderInvalidEmailError:
      summary: 電子郵件格式錯誤
      value:
        success: false
        error:
          code: "VALIDATION_ERROR"
          message: "電子郵件格式不正確"
          details:
            - field: "parent.email"
              reason: "must be a valid email address"
          traceId: "ord123def002"
        timestamp: "2026-01-26T10:30:00+08:00"
    OrderInvalidPhoneError:
      summary: 電話格式錯誤
      value:
        success: false
        error:
          code: "VALIDATION_ERROR"
          message: "電話格式不正確"
          details:
            - field: "parent.phone"
              reason: "must match pattern ^09\\d{8}$"
          traceId: "ord123def003"
        timestamp: "2026-01-26T10:30:00+08:00"
    OrderAgeNotEligibleError:
      summary: 小孩年齡不符合賽事限制
      value:
        success: false
        error:
          code: "AGE_NOT_ELIGIBLE"
          message: "小孩年齡不符合賽事年齡限制"
          details:
            - field: "child.birthDate"
              reason: "child age (6) is not within event age range (8-12)"
          traceId: "ord123def004"
        timestamp: "2026-01-26T10:30:00+08:00"
    OrderEventFullError:
      summary: 賽事名額已滿
      value:
        success: false
        error:
          code: "EVENT_FULL"
          message: "賽事名額已滿，無法報名"
          details:
            - field: "eventId"
              reason: "event has reached maximum capacity (30/30)"
          traceId: "ord123def005"
        timestamp: "2026-01-26T10:30:00+08:00"
    OrderRegistrationClosedError:
      summary: 報名已截止
      value:
        success: false
        error:
          code: "REGISTRATION_CLOSED"
          message: "賽事報名已截止"
          details:
            - field: "eventId"
              reason: "registration deadline was 2026-01-20T23:59:59+08:00"
          traceId: "ord123def006"
        timestamp: "2026-01-26T10:30:00+08:00"
    OrderDuplicateError:
      summary: 重複報名
      value:
        success: false
        error:
          code: "DUPLICATE_REGISTRATION"
          message: "此小孩已報名該賽事"
          details:
            - field: "child"
              reason: "child has already registered for this event"
          traceId: "ord123def007"
        timestamp: "2026-01-26T10:30:00+08:00"
    OrderInvalidBirthDateError:
      summary: 出生日期為未來日期
      value:
        success: false
        error:
          code: "VALIDATION_ERROR"
          message: "出生日期不可為未來日期"
          details:
            - field: "child.birthDate"
              reason: "birthDate must be in the past, got: 2028-06-01"
          traceId: "ord123def011"
        timestamp: "2026-01-26T10:30:00+08:00"
    OrderNonStrictAgeSuccess:
      summary: 彈性年齡驗證模式下年齡略超出範圍但允許報名（含 warnings）
      value:
        success: true
        data:
          id: "660e8400-e29b-41d4-a716-446655440001"
          eventId: "550e8400-e29b-41d4-a716-446655440002"
          eventName: "2026 春季兒童籃球體驗營（彈性收費）"
          status: "PENDING"
          parent:
            name: "陳大文"
            email: "chen@example.com"
            phone: "0933112233"
          child:
            name: "陳小強"
            birthDate: "2012-11-05"
            gender: "MALE"
          childAgeAtEvent: 13
          fee: 1500
          paymentDeadline: "2026-02-01T23:59:59+08:00"
          createdAt: "2026-01-26T10:30:00+08:00"
          updatedAt: "2026-01-26T10:30:00+08:00"
          warnings:
            - code: "AGE_BOUNDARY_WARNING"
              message: "小孩年齡（13歲）超出賽事建議年齡範圍（8-12歲），已依主辦單位彈性規定允許報名，請確認參加意願"
        timestamp: "2026-01-26T10:30:00+08:00"

    # ===== Admin Examples =====
    CreateEventRequestExample:
      summary: 建立賽事請求範例
      value:
        name: "2026 春季兒童籃球營"
        description: "專為 8-12 歲兒童設計的籃球訓練營，由專業教練指導基礎籃球技巧，包含運球、傳球、投籃等基礎訓練。活動結束後將舉辦小型友誼賽，讓孩子們在比賽中學習團隊合作。"
        ageMin: 8
        ageMax: 12
        ageRestrictionNote: "適合 8 至 12 歲兒童參加"
        strictAgeEnforcement: true
        startTime: "2026-03-15T09:00:00+08:00"
        endTime: "2026-03-15T12:00:00+08:00"
        location: "台北市大安運動中心"
        address: "台北市大安區辛亥路三段55號"
        capacity: 30
        registrationDeadline: "2026-03-10T23:59:59+08:00"
        fee: 1500
        organizer: "台北市籃球協會"
        contactEmail: "events@basketball.org.tw"
        contactPhone: "02-1234-5678"
...
    AdminEventValidationError:
      summary: 賽事參數驗證失敗
      value:
        success: false
        error:
          code: "VALIDATION_ERROR"
          message: "請求參數驗證失敗"
          details:
            - field: "ageMin"
              reason: "must be less than or equal to ageMax"
            - field: "startTime"
              reason: "must be before endTime"
            - field: "registrationDeadline"
              reason: "must be before startTime"
          traceId: "adm123def001"
        timestamp: "2026-01-26T10:30:00+08:00"
    UnauthorizedError:
      summary: 未認證
      value:
        success: false
        error:
          code: "UNAUTHORIZED"
          message: "未提供有效的認證憑證"
          traceId: "auth123def001"
        timestamp: "2026-01-26T10:30:00+08:00"
    ForbiddenError:
      summary: 權限不足
      value:
        success: false
        error:
          code: "FORBIDDEN"
          message: "您沒有權限執行此操作（需要管理員權限）"
          traceId: "auth123def002"
        timestamp: "2026-01-26T10:30:00+08:00"